Introduce async/dynamic GLTF factory
