+++ packages/dev/loaders/src/dynamic.ts
@@ -1,5 +1,7 @@
 import { registerOBJLoader } from "./OBJ/objDynamicFileLoader";
+import { registerGLTFLoader } from "./glTF/glTFDynamicFileLoader";
 
 export function registerLoaders() {
     registerOBJLoader();
+    registerGLTFLoader();
 }
